Add files via upload
Project Name - Bike Sharing Demand Prediction
Problem Statement
BUSINESS PROBLEM OVERVIEW

The dataset is of Seoul's public bike rental service called "Ttareungi," also known as Seoul Bike. It has become one of the capital city's most popular public transport systems, having been used more than 100 million times since the service was launched in December 2015, as reported in early 2022. The popularity however has posed the challenge of keeping the bikes available as and when the demand for them rises.

So my goal is to identify the variables that have a say in the demand of the bikes and then build a model that helps predict the demands as and when the climatic conditions change.