Add files via upload
This presentation provides an overview of the Seoul Bike Sharing project, including the dataset analysis, insights derived through exploratory data analysis (EDA), and the development of a predictive model. It also highlights the model deployment using Gradio, showcasing the prediction of real-time bike demand. For a deeper dive into the methodology and results, please refer to the full presentation.